Supresss compiler warning - implicit conversion loses integer precisi…

+++ Objective-Shorthand/Objective-Shorthand/NSArray+FunctionalMethods.m
@@ -57,7 +57,7 @@
     if (self.count == 0) {
         return nil;
     }
-    NSUInteger randomIndex = arc4random_uniform(self.count);
+    NSUInteger randomIndex = arc4random_uniform((u_int32_t)self.count);
     return self[randomIndex];
 }
